refactor: Move PurposeEditDialogs and SectionEditDialogs in EditDialogs
