+++ src/components/EditDialogs/SectionEditDialog/helpers.js
